Convolutions

TinOpener.convolve — Function

convolve(x::Vector{Float64}, t::Int, K::Function, S::Int)

Obtain a discrete convolution of vector x with convolution kernel K at time/index t, up to S times in the past.

In other words, compute the quantity

\[(K \star x)_t = \sum_{s=0}^S K(s) x(t-s).\]

TinOpener.weak_kernel — Function

weak_kernel(s; T = 1.0)

The "weak" (exponentially decaying) convolution kernel

\[K(s) = \frac{1}{T} e^{-s/T}.\]

Reference

Cushing FIXME

TinOpener.strong_kernel — Function

strong_kernel(s; T = 1.0)

The "strong" convolution kernel

\[K(s) = \frac{s}{T^2} e^{-s/T}.\]

This has maximum at $s = T$.

Reference

Cushing FIXME

TinOpener.zc_kernel — Function

zc_kernel(s; T = 1.0, U = 2.0)

The "zero-crossing" kernel

\[K(s) = (U - s) \frac{s}{T^2} e^{-s/T}.\]

TinOpener.bzc_kernel — Function

bzc_kernel(s; T = 1.0)

The "balanced zero-crossing" kernel

\[K(s) = (2T - s) \frac{s}{T^2} e^{-s/T}.\]

This places equal mass on each side of the crossover point.