Distributions

TinOpener.BiBeta — Type

BiBeta

A bivariate beta distribution.

The Olkin–Trikalinos construction for a bivariate beta distribution, with parameters α11, α10, α01 and α00.

Details

The construction is based on a Dirichlet distribution. Suppose $(U_{11}, U_{10}, U_{01}, U_{00}) \sim \textnormal{Dirichlet}(\alpha)$ where $\alpha = (\alpha_{11}, \alpha_{10}, \alpha_{01}, \alpha_{00})$. Define $X = U_{11} + U_{10}$ and $Y = U_{11} + U_{01}$. Then the joint density of $X$ and $Y$ is

\[f(x,y) = B(\alpha)^{-1} \int_{\Omega} u_{11}^{\alpha_{11} - 1} (x - u_{11})^{\alpha_{10} - 1} (y - u_{11})^{\alpha_{01} - 1} (1 - x - y + u_{11})^{\alpha_{00} - 1} du_{11}\]

with $\Omega = \{u_{11} : \max(0, x+y-1) < u_{11} < \min(x,y)\}$ and $B(\alpha)$ the generalized beta function $B(\alpha) = \Gamma(\alpha_{11}) \Gamma(\alpha_{10}) \Gamma(\alpha_{01}) \Gamma(\alpha_{00}) / \Gamma(\alpha_{11} + \alpha_{10} + \alpha_{01} + \alpha_{00})$.

Reference

Olkin, I. & Trikalinos, T. A. (2015) Constructions for a bivariate beta distributions. Statistics and Probability Letters, 96, 54–60. http://dx.doi.org/10.1016/j.spl.2014.09.013.

TinOpener.pdf — Function

pdf(d::BiBeta, x::Vector{Float64})

Density of the bivariate beta distribution d, evaluated at point x.

The integral in the density is solved numerically using FIXME.

TinOpener.logpdf — Function

logpdf(d::BiBeta, x::Vector{Float64})

Logarithmic density of the bivariate beta distribution d, evaluated at point x.

The integral in the density is solved numerically using FIXME.

TinOpener.momentmatch — Function

momentmatch(μ1, μ2, var1, var2, ρ)

Find a bivariate beta distribution (BiBeta), i.e. its parameters, by moment-matching.

Arguments

μ1: mean of first covariate

μ2: mean of second covariate

var1: variance of first covariate

var2: variance of second covariate

ρ: correlation between covariates

Value

A NamedTuple with the following fields:

d: distribution found

sol: solution object from numerical solver

Details

The best-fitting parameter vector $\alpha$ is found by numerical optimization as per the procedure outlined by Olkin & Trikalinos (2015). The Nelder–Mead algorithm is used.

Reference

Olkin, I. & Trikalinos, T. A. (2015) Constructions for a bivariate beta distributions. Statistics and Probability Letters, 96, 54–60. http://dx.doi.org/10.1016/j.spl.2014.09.013.