Kalman filtering

Methods

TinOpener.forward! — Function

forward!(kr::InvariantKalmanRecursion,
         y::Vector{Float64})

Update a Kalman recursion with one data point y.

forward!(kr::KalmanRecursion,
         y::Vector{Float64})

Update a Kalman recursion with one data point y.

TinOpener.get_filtered_states — Function

get_filtered_states(kr::AbstractKalmanRecursion)

Get the filtered states from a Kalman recursion.

TinOpener.get_filtered_variances — Function

get_filtered_variances(kr::AbstractKalmanRecursion)

Get the filtered variances from a Kalman recursion.

TinOpener.get_smoothed_states — Function

get_smoothed_states(kr::AbstractKalmanRecursion)

Get the smoothed states from a Kalman recursion.

TinOpener.get_smoothed_variances — Function

get_smoothed_variances(kr::AbstractKalmanRecursion)

Get the smoothed variances from a Kalman recursion.

TinOpener.kalman_recursion — Function

kalman_recursion(model::AbstractInvariantLinearGaussianSSM,
                 a::Vector{Float64},
                 P::Matrix{Float64})

Initialize a Kalman recursion for any descendant of AbstractInvariantLinearGaussianSSM, with (known/assumed) initial state mean a and variance P.

Value

An InvariantKalmanRecursion object.

kalman_recursion(model::AbstractLinearGaussianSSM,
                 a::Vector{Float64},
                 P::Matrix{Float64})

Initialize a Kalman recursion for any descendant of AbstractLinearGaussianSSM, with (known/assumed) initial state mean a and variance P.

Value

A KalmanRecursion object.

TinOpener.loglikelihood — Function

loglikelihood(kr::AbstractKalmanRecursion)

Obtain the log-likelihood from a Kalman filter.

TinOpener.smooth! — Function

smooth!(kr::InvariantKalmanRecursion,
        y::Vector{Vector{Float64}})

Kalman smoother.

smooth!(kr::KalmanRecursion,
        y::Vector{Vector{Float64}})

Kalman smoother.

Types

TinOpener.InvariantKalmanRecursion — Type

InvariantKalmanRecursion <: AbstractKalmanRecursion

A struct to hold the matrices and vectors required for Kalman filtering and smoothing in the case the model matrices are time-invariant.

Fields

For the meaning of the fields of this struct, consult Durbin & Koopman (2012), particularly the filtering and smoothing recursion equations (4.24) and (4.44) and their surrounding text.

Z – the observation matrix

T – the process matrix

R – process noise matrix

H – covariances of observation noise

Q – covariances of process noise

a – state means

v – difference between observation and expected observation (i.e. forecast error, i.e. "innovation")

F – variance of v

P – state variances

K – Kalman gain

αhat – smoothed state mean

r – weighted sum of innovations, used in the computation of αhat

N – used in the computation of V

V – smoothed state variance

Reference

Durbin, J. & Koopman, S. J. (2012) Time series analysis by state space methods. 2nd edition. Oxford: Oxford University Press.

Abstract types

TinOpener.AbstractKalmanRecursion — Type

AbstractKalmanRecursion

Abstract type that is the supertype of various Kalman recursion structs.