Parallel processing

TinOpener.mapply — Function

mapply(fun, vars)

Multidimensional apply: apply function fun to each combination of the variable values in vars, a NamedTuple.

Value

Returns a named tuple with fields:

grid: grid of values, a DataFrame

result: vector of results, i.e. the output of fun when applied to each row of grid

Example

using Distributed

addprocs(6)

@everywhere myfun(x) = sum(x)

mapply(myfun, (var1 = 1:3, var2 = 1:10, var3 = 10:20))