Reference

TinOpener.bootstrap — Function

bootstrap(x::DataFrame)

Bootstrap a DataFrame, i.e. sample its rows with replacement.

bootstrap(x::GroupedDataFrame{DataFrame})

Bootstrap a GroupedDataFrame, i.e. separately call bootstrap on each of its subdataframes.

TinOpener.bin_time — Function

bin_time(x::DataFrame; time_variable = "date", bin_resolution = 10)

Bin the time variable in a DataFrame into bin_resolution equally spaced bins. Returns a BinnedCorpus.

TinOpener.categorical_to_proportion — Function

categorical_to_proportion(x::BinnedCorpus, variable, ref_level;
    expand_missing = true, shuffle = false)

Transform a categorical variable into its proportion, calculated per time bin.

The variable is first transformed into a vector of ones and zeroes; whenever the value coincides with ref_level, we replace the value with a one, otherwise we replace it with a zero. The proportion of ref_level per time bin is then calculated as the mean over this vector of ones and zeroes.

The output is a BinnedTimeSeries.

TinOpener.impute — Function

impute(x::BinnedTimeSeries; nnsize = 2)

Carry out nearest-neighbour imputation (with neighbourhood size nnsize) over a BinnedTimeSeries.

Return value is a BinnedTimeSeries with missing values replaced by imputed values.