State-space models

Methods

TinOpener.simulate — Function

simulate(model::AbstractInvariantLinearGaussianSSM;
         iter::Int,
         dist0)

Fimulate an AbstractInvariantLinearGaussianSSM for iter time steps from an initial state sampled from distribution dist0.

Returns the simulated series of states, along with simulated series of observations.

Types and constructors

Abstract types

TinOpener.AbstractInvariantLinearGaussianSSM — Type

AbstractInvariantLinearGaussianSSM

Represents a linear Gaussian state space model with time-invariant matrices.

TinOpener.AbstractLinearGaussianSSM — Type

AbstractLinearGaussianSSM

TinOpener.KalmanRecursion — Type

KalmanRecursion <: AbstractKalmanRecursion

A struct to hold the matrices and vectors required for Kalman filtering and smoothing in the case the model matrices are time-invariant.

Fields

For the meaning of the fields of this struct, consult Durbin & Koopman (2012), particularly the filtering and smoothing recursion equations (4.24) and (4.44) and their surrounding text.

Z – the observation matrix

T – the process matrix

R – process noise matrix

H – covariances of observation noise

Q – covariances of process noise

a – state means

v – difference between observation and expected observation (i.e. forecast error, i.e. "innovation")

F – variance of v

P – state variances

K – Kalman gain

αhat – smoothed state mean

r – weighted sum of innovations, used in the computation of αhat

N – used in the computation of V

V – smoothed state variance

t – current time

Reference

Durbin, J. & Koopman, S. J. (2012) Time series analysis by state space methods. 2nd edition. Oxford: Oxford University Press.