General time series tools

Methods

TinOpener.bin_time — Function

bin_time(data::DataFrame; time_variable = "date", bin_resolution = 10)

Bin the time variable in a DataFrame into bin_resolution equally spaced bins.

The left endpoint of the first bin is the minimum of time_variable; the right endpoint of the last bin is the maximum of time_variable.

Value

A BinnedData object.

TinOpener.binary_to_proportion — Function

binary_to_proportion(x::BinnedData, variable, ref_level;
                     shuffle = false, α = 0)

Transforms a binary categorical variable into its proportion, calculated per time bin.

The variable is first transformed into a vector of ones and zeroes; whenever the value coincides with ref_level, we replace the value with a one, otherwise we replace it with a zero. The proportion of ref_level per time bin is then calculated as the mean over this vector of ones and zeroes.

If shuffle is true, a bootstrap (random sample with replacement) is taken per each time bin.

The keyword argument α can be used to perform additive smoothing. For instance, setting α = 1 gives add-one (Laplace) smoothing.

Value

The output is a BinnedTimeSeries.

TinOpener.bootstrap — Function

bootstrap(x::DataFrame)

Bootstrap a DataFrame, i.e. sample its rows with replacement.

bootstrap(x::SubDataFrame)

Bootstrap a SubDataFrame, i.e. sample its rows with replacement.

bootstrap(x::GroupedDataFrame{DataFrame})

Bootstrap a GroupedDataFrame, i.e. separately call bootstrap on each of its subdataframes.

TinOpener.expand_missing — Function

expand_missing(x::BinnedTimeSeries)

Expands the data in a BinnedTimeSeries so that time bins with no data are explicitly flagged as having missing proportions.

TinOpener.shuffle — Function

shuffle(x::BinnedTimeSeries)

Return a random permutation of a BinnedTimeSeries. Mainly useful for surrogate testing purposes.

shuffle(x::Vector{Float64})

Return a random permutation of an array of floats. Mainly useful for surrogate testing purposes.

Types

TinOpener.BinnedData — Type

BinnedData

A data structure containing a time-binned dataframe. Fields:

data: the dataframe

bin_resolution: binning resolution (see bin_time)

midpoints: midpoints of the time bins on the original time scale

TinOpener.BinnedTimeSeries — Type

BinnedTimeSeries

A data structure containing a time series produced by binning a dataframe. Fields:

data: the time series

bin_resolution: binning resolution (see bin_time)

midpoints: midpoints of the time bins on the original time scale