State-space models

Methods

TinOpener.simulate — Function

simulate(model::AbstractInvariantLinearGaussianSSM;
         iter::Int,
         dist0)

Fimulate an AbstractInvariantLinearGaussianSSM for iter time steps from an initial state sampled from distribution dist0.

Returns the simulated series of states, along with simulated series of observations.

Types and constructors

TinOpener.MomentumCoupledSLSSM — Type

MomentumCoupledSLSSM <: AbstractInvariantLinearGaussianSSM

A momentum-coupled stochastic logistic state space model.

Fields

Z

T

R

H

Q

For the meaning of these, see InvariantKalmanRecursion.

TinOpener.MomentumCoupledSLSSM — Method

MomentumCoupledSLSSM(κ:Float64,
                     κp::Float64,
                     σ2y::Float64,
                     σ2z::Float64,
                     σ2s::Float64,
                     qy::Float64,
                     qz::Float64,
                     qs::Float64)

Constructor for a MomentumCoupledSLSSM.

TinOpener.SLSSM — Type

SLSSM <: AbstractInvariantLinearGaussianSSM

A one-dimensional stochastic logistic state space model.

Fields

Z

T

R

H

Q

For the meaning of these, see InvariantKalmanRecursion.

TinOpener.SLSSM — Method

SLSSM(σ2z::Float64,
      σ2s::Float64,
      σ2y::Float64)

Constructor for an SLSSM.

Abstract types

TinOpener.AbstractInvariantLinearGaussianSSM — Type

AbstractInvariantLinearGaussianSSM

Represents a linear Gaussian state space model with time-invariant matrices.